backend/requirements.txt
# Core
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Scraping
praw>=7.7.1
youtube-transcript-api>=0.6.1
playwright>=1.40.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# AI
openai>=1.10.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langgraph>=0.0.20

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Code Quality
mypy>=1.8.0
black>=24.1.0
